refac load_users
